Fix bugs in CExpRKMethod.cpp: change size_t mRootId to int size_t mRootId, add pointer check before allocate new space
	modified:   CExpRKMethod.cpp
	modified:   CExpRKMethod.h

+++ CExpRKMethod.h
@@ -228,7 +228,7 @@ class CExpRKMethod
    * mRootId, a size_t variable indicating which root is found
    *          0 <= mRootId <= mRootNum-1
    */
-  size_t mRootId;
+  int mRootId;
 
 
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
